Fix BrowserSync UI
The UI only runs from an IP; it won't work with a URL.

Remove the changes to Traefik.

Update Traefik.

Add a boot script to Info and use it to get the IP. We may need CGI
later, but this is the only non-static part of the page thus far.